review-fix: Move SentModal into fether-react so MIN_CONFIRMATIONS not passed as prop
